myproject/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto==2.49.0
boto3==1.28.59
botocore==1.31.59
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
Django==4.1.12
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.4.2
geographiclib==2.0
geolocator==0.1.1
geopy==2.4.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
numpy==1.26.0
outcome==1.2.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pycparser==2.21
pymongo==4.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytesseract==0.3.10
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.2.4
tabulate==0.9.0
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.17
utc==0.0.3
webdriver-manager==4.0.1
wsproto==1.2.0
openpyxl==3.1.2
djangorestframework
django-cors-headers